+++ src/main/java/com/banquito/core/banking/seguridadbanco/dao/PersonalBancarioRepository.java
@@ -2,7 +2,6 @@ package com.banquito.core.banking.seguridadbanco.dao;
 
 import java.util.List;
 import java.util.Optional;
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import org.springframework.data.repository.CrudRepository;
@@ -11,7 +10,7 @@ import org.springframework.stereotype.Repository;
 
 import com.banquito.core.banking.seguridadbanco.domain.PersonalBancario;
 @Repository
-public interface PersonalBancarioRepository extends CrudRepository<PersonalBancario, BigDecimal> {
+public interface PersonalBancarioRepository extends CrudRepository<PersonalBancario, Integer> {
     // public List<PersonalBancario> findByUsuarioAndCodRol(String usuario, BigDecimal codRol);
 
     public List<PersonalBancario> findByFechaCreacionBetween(Timestamp fechaInicio, Timestamp fechaFin);
