+++ src/main/java/com/banquito/core/banking/seguridadbanco/services/AccesoPbRolService.java
@@ -1,6 +1,6 @@
 package com.banquito.core.banking.seguridadbanco.services;
 
-import java.math.BigDecimal;
+
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
@@ -21,7 +21,7 @@ public class AccesoPbRolService {
         this.accesoPbRolRepository = accesoPbRolRepository;
     }
 
-    public Optional<AccesoPbRol> getById(BigDecimal codRol, BigDecimal codPerBan) {
+    public Optional<AccesoPbRol> getById(Integer codRol, Integer codPerBan) {
         AccesoPbRolPK accesoPbRolPK = new AccesoPbRolPK(codRol, codPerBan);
         return this.accesoPbRolRepository.findById(accesoPbRolPK);
     }
@@ -35,7 +35,7 @@ public class AccesoPbRolService {
         }
     }
 
-    public void delete(BigDecimal codRol, BigDecimal codPerBan) {
+    public void delete(Integer codRol, Integer codPerBan) {
         try {
             Optional<AccesoPbRol> accesoPbRol = getById(codRol, codPerBan);
             if (accesoPbRol.isPresent()) {
